Fixed Audit Task not updating tags
If you have a previous installation of ExHentai-Archive, you will need
to at a lastaudit column to the gallery table. The column needs to be a
datetime type that defaults to NULL. This will take a very long time on
systems with large amounts of archives. This operation is resumable as
well.